turned from executable to library

+++ source/rgb.hpp
@@ -1,15 +1,13 @@
 #ifndef RGB_HPP
 #define RGB_HPP
 
-#include <cstdint>
-
 struct RGB {
-  uint8_t R{};
-  uint8_t G{};
-  uint8_t B{};
+  int R{};
+  int G{};
+  int B{};
 
   RGB() = default;
-  RGB(uint8_t, uint8_t, uint8_t);
+  RGB(int, int, int);
 };
 
 #endif
